Move functionalities under lib
